linux support for scripts
