+++ bin/umodel/win32/umodel.cfg
@@ -1,0 +1,19 @@
+Export =
+{
+    ExportPath = "C:\Users\prosser\Downloads\umodel_win32\UmodelExport"
+    ExportDdsTexture = false
+    SkeletalMeshFormat = 0
+    StaticMeshFormat = 0
+    TextureFormat = 0
+    ExportMeshLods = false
+    SaveUncooked = false
+    SaveGroups = false
+    DontOverwriteFiles = false
+}
+SavePackages =
+{
+    SavePath = "C:/Users/prosser/Downloads/umodel_win32/UmodelSaved"
+    KeepDirectoryStructure = true
+}
+bShowExportOptions = true
+bShowSaveOptions = true
